refactor(keyvox-tts-runtime): split PocketTTS runtime into focused runtime components
- replace `Packages/KeyVoxTTS/Sources/KeyVoxTTS/KeyVoxPocketTTSRuntime.swift` with a dedicated `Packages/KeyVoxTTS/Sources/KeyVoxTTS/KeyVoxPocketTTSRuntime/` directory
- keep `KeyVoxPocketTTSRuntime.swift` focused on the public runtime actor, compute-mode entrypoints, and stream setup
- move asset validation, model loading, constants loading, and cached voice and KV snapshot loading into `KeyVoxPocketTTSRuntime+Assets.swift`
- move chunk planning, text prefill, frame generation, batching, and stream emission into `KeyVoxPocketTTSStreamGenerator.swift`
- move compute-mode state ownership and `KeyVoxTTSComputeMode` logging helpers into `KeyVoxPocketTTSComputeModeController.swift`
- preserve the existing fast-mode batching and deterministic remaining-sample emission behavior while separating runtime orchestration from stream generation and asset-loading concerns
- relax the split runtime’s internal visibility just enough for the new sibling files to share model, cache, and compute-mode state cleanly without keeping the old monolith together
- run `swift test --package-path "/Users/domesposito/Projects/KeyVox/Packages/KeyVoxTTS"` after the split; all 7 tests passed, with the same existing `.build/build.db` disk I/O error still emitted by SwiftPM on exit
