+++ Packages/KeyVoxTTS/Sources/KeyVoxTTS/KeyVoxPocketTTSRuntime/KeyVoxPocketTTSStreamGenerator.swift
@@ -1,207 +1,7 @@
 @preconcurrency import CoreML
 import Foundation
 
-public enum KeyVoxTTSComputeMode: Sendable {
-    case foreground
-    case backgroundSafe
-}
-
-public actor KeyVoxPocketTTSRuntime {
-    fileprivate struct ModelSet {
-        let condStepModel: MLModel
-        let flowLanguageModel: MLModel
-        let flowDecoderModel: MLModel
-        let mimiDecoderModel: MLModel
-    }
-
-    private let assetLayout: KeyVoxTTSAssetLayout
-    private var foregroundModels: ModelSet?
-    private var backgroundModels: ModelSet?
-    private var constantsBundle: PocketTTSConstantsBundle?
-    private var cachedVoices: [KeyVoxTTSVoice: PocketTTSVoiceConditioning] = [:]
-    private var cachedVoiceKVSnapshots: [KeyVoxTTSVoice: PocketTTSInferenceTypes.KVCacheState] = [:]
-    private let computeModeController = ComputeModeController()
-
-    public init(assetLayout: KeyVoxTTSAssetLayout) {
-        self.assetLayout = assetLayout
-    }
-
-    public func prepareIfNeeded() async throws {
-        guard foregroundModels == nil || backgroundModels == nil else { return }
-
-        try validateAssetLayout()
-        let loadStart = CFAbsoluteTimeGetCurrent()
-
-        let newForegroundModels = try loadModelSet(computeUnits: .cpuAndGPU)
-        Self.log("Loaded foreground PocketTTS models with cpuAndGPU in \(String(format: "%.3f", CFAbsoluteTimeGetCurrent() - loadStart))s")
-
-        let backgroundLoadStart = CFAbsoluteTimeGetCurrent()
-        let newBackgroundModels = try loadModelSet(computeUnits: .cpuOnly)
-        Self.log("Loaded background PocketTTS models with cpuOnly in \(String(format: "%.3f", CFAbsoluteTimeGetCurrent() - backgroundLoadStart))s")
-        
-        let newConstantsBundle = try PocketTTSAssetLoader.loadConstants(from: assetLayout)
-        
-        foregroundModels = newForegroundModels
-        backgroundModels = newBackgroundModels
-        constantsBundle = newConstantsBundle
-    }
-
-    public func setPreferredComputeMode(_ mode: KeyVoxTTSComputeMode) async {
-        await computeModeController.setMode(mode)
-        Self.log("Preferred compute mode set to \(mode.logName).")
-    }
-
-    public func synthesizeStreaming(
-        text: String,
-        voice: KeyVoxTTSVoice,
-        fastModeEnabled: Bool = false,
-        seed: UInt64? = nil
-    ) async throws -> AsyncThrowingStream<KeyVoxTTSAudioFrame, Error> {
-        try await prepareIfNeeded()
-
-        let trimmedText = text.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmedText.isEmpty else {
-            throw KeyVoxTTSError.inferenceFailure("PocketTTS requires non-empty text.")
-        }
-        Self.log("Starting synthesis for text length \(trimmedText.count) chars.")
-
-        let constants = try constants()
-        let voiceConditioning = try loadVoiceConditioning(voice)
-        let voicePrefillStart = CFAbsoluteTimeGetCurrent()
-        let currentMode = await computeModeController.mode()
-        let prefillModel = currentMode == .backgroundSafe 
-            ? try modelSet(backgroundModels, modeName: "background").condStepModel
-            : try modelSet(foregroundModels, modeName: "foreground").condStepModel
-        let voiceKVSnapshot = try await loadVoiceKVSnapshot(
-            for: voice,
-            voiceConditioning: voiceConditioning,
-            model: prefillModel
-        )
-        let chunks = PocketTTSChunkPlanner.chunk(
-            trimmedText,
-            tokenizer: constants.tokenizer,
-            fastModeEnabled: fastModeEnabled
-        )
-        Self.log(
-            "Voice prefill completed in \(String(format: "%.3f", CFAbsoluteTimeGetCurrent() - voicePrefillStart))s. Chunks: \(chunks.count)"
-        )
-        let generator = try StreamGenerator(
-            chunks: chunks,
-            fastModeEnabled: fastModeEnabled,
-            constants: constants,
-            voiceConditioning: voiceConditioning,
-            foregroundModels: try modelSet(foregroundModels, modeName: "foreground"),
-            backgroundModels: try modelSet(backgroundModels, modeName: "background"),
-            initialMimiState: try PocketTTSMimiInference.initialState(from: assetLayout),
-            beginningOfSequenceEmbedding: try PocketTTSInferenceUtilities.createBeginningOfSequenceEmbedding(constants.beginningOfSequenceEmbedding),
-            voiceKVSnapshot: voiceKVSnapshot,
-            computeModeController: computeModeController,
-            seed: seed ?? UInt64.random(in: 0...UInt64.max)
-        )
-
-        return AsyncThrowingStream { continuation in
-            let task = Task {
-                await generator.generate(into: continuation)
-            }
-            continuation.onTermination = { _ in
-                task.cancel()
-            }
-        }
-    }
-
-    private func validateAssetLayout() throws {
-        let requiredModelNames = [
-            PocketTTSConstants.ModelName.condStep,
-            PocketTTSConstants.ModelName.flowLMStep,
-            PocketTTSConstants.ModelName.flowDecoder,
-            PocketTTSConstants.ModelName.mimiDecoder,
-        ]
-
-        for modelName in requiredModelNames {
-            let modelURL = assetLayout.compiledModelURL(named: modelName)
-            guard FileManager.default.fileExists(atPath: modelURL.path) else {
-                throw KeyVoxTTSError.missingModel("PocketTTS model \(modelName) is missing.")
-            }
-        }
-
-        guard FileManager.default.fileExists(atPath: assetLayout.constantsDirectoryURL.path) else {
-            throw KeyVoxTTSError.invalidAssetLayout("PocketTTS constants directory is missing.")
-        }
-    }
-
-    private func constants() throws -> PocketTTSConstantsBundle {
-        guard let constantsBundle else {
-            throw KeyVoxTTSError.missingAsset("PocketTTS constants are not loaded.")
-        }
-        return constantsBundle
-    }
-
-    private func loadVoiceConditioning(_ voice: KeyVoxTTSVoice) throws -> PocketTTSVoiceConditioning {
-        if let cached = cachedVoices[voice] {
-            return cached
-        }
-
-        let loaded = try PocketTTSAssetLoader.loadVoice(voice, from: assetLayout)
-        cachedVoices[voice] = loaded
-        return loaded
-    }
-
-    private func loadVoiceKVSnapshot(
-        for voice: KeyVoxTTSVoice,
-        voiceConditioning: PocketTTSVoiceConditioning,
-        model: MLModel
-    ) async throws -> PocketTTSInferenceTypes.KVCacheState {
-        if let cachedSnapshot = cachedVoiceKVSnapshots[voice] {
-            Self.log("Reusing cached voice prefill for \(voice.rawValue).")
-            return try PocketTTSKVCacheInference.cloneState(cachedSnapshot)
-        }
-
-        let snapshot = try await PocketTTSKVCacheInference.prefillVoice(
-            voice: voiceConditioning,
-            model: model
-        )
-        cachedVoiceKVSnapshots[voice] = snapshot
-        Self.log("Cached voice prefill for \(voice.rawValue).")
-        return try PocketTTSKVCacheInference.cloneState(snapshot)
-    }
-
-    private func loadModelSet(computeUnits: MLComputeUnits) throws -> ModelSet {
-        let configuration = MLModelConfiguration()
-        configuration.computeUnits = computeUnits
-
-        return ModelSet(
-            condStepModel: try MLModel(
-                contentsOf: assetLayout.compiledModelURL(named: PocketTTSConstants.ModelName.condStep),
-                configuration: configuration
-            ),
-            flowLanguageModel: try MLModel(
-                contentsOf: assetLayout.compiledModelURL(named: PocketTTSConstants.ModelName.flowLMStep),
-                configuration: configuration
-            ),
-            flowDecoderModel: try MLModel(
-                contentsOf: assetLayout.compiledModelURL(named: PocketTTSConstants.ModelName.flowDecoder),
-                configuration: configuration
-            ),
-            mimiDecoderModel: try MLModel(
-                contentsOf: assetLayout.compiledModelURL(named: PocketTTSConstants.ModelName.mimiDecoder),
-                configuration: configuration
-            )
-        )
-    }
-
-    private func modelSet(_ modelSet: ModelSet?, modeName: String) throws -> ModelSet {
-        guard let modelSet else {
-            throw KeyVoxTTSError.missingModel("PocketTTS \(modeName) model set is not loaded.")
-        }
-        return modelSet
-    }
-
-    private static func log(_ message: String) {
-        NSLog("[KeyVoxPocketTTSRuntime] %@", message)
-    }
-}
-
-private actor StreamGenerator {
+actor KeyVoxPocketTTSStreamGenerator {
     private enum StreamingPolicy {
         static let defaultBatchedFrameCount = 8
     }
@@ -527,28 +327,5 @@ private actor StreamGenerator {
 
     private static func log(_ message: String) {
         NSLog("[KeyVoxPocketTTSStream] %@", message)
-    }
-}
-
-private actor ComputeModeController {
-    private var preferredMode: KeyVoxTTSComputeMode = .foreground
-
-    func setMode(_ mode: KeyVoxTTSComputeMode) {
-        preferredMode = mode
-    }
-
-    func mode() -> KeyVoxTTSComputeMode {
-        preferredMode
-    }
-}
-
-private extension KeyVoxTTSComputeMode {
-    var logName: String {
-        switch self {
-        case .foreground:
-            return "foreground"
-        case .backgroundSafe:
-            return "backgroundSafe"
-        }
     }
 }
